+++ source/client/client_worker_impl.cc
@@ -11,6 +11,7 @@ namespace Client {
 ClientWorkerImpl::ClientWorkerImpl(Envoy::Api::Api& api, Envoy::ThreadLocal::Instance& tls,
                                    Envoy::Upstream::ClusterManagerPtr& cluster_manager,
                                    const BenchmarkClientFactory& benchmark_client_factory,
+                                   const TerminationPredicateFactory& termination_predicate_factory,
                                    const SequencerFactory& sequencer_factory,
                                    const HeaderSourceFactory& header_generator_factory,
                                    Envoy::Stats::Store& store, const int worker_number,
@@ -24,8 +25,11 @@ ClientWorkerImpl::ClientWorkerImpl(Envoy::Api::Api& api, Envoy::ThreadLocal::Ins
       benchmark_client_(benchmark_client_factory.create(
           api, *dispatcher_, *worker_number_scope_, cluster_manager, http_tracer_,
           fmt::format("{}", worker_number), *header_generator_)),
-      sequencer_(
-          sequencer_factory.create(time_source_, *dispatcher_, starting_time, *benchmark_client_)),
+      termination_predicate_(
+          termination_predicate_factory.create(time_source_, *worker_number_scope_, starting_time)),
+      sequencer_(sequencer_factory.create(time_source_, *dispatcher_, starting_time,
+                                          *benchmark_client_, *termination_predicate_,
+                                          *worker_number_scope_)),
       prefetch_connections_(prefetch_connections) {}
 
 void ClientWorkerImpl::simpleWarmup() {
@@ -46,22 +50,26 @@ void ClientWorkerImpl::work() {
   benchmark_client_->setMeasureLatencies(true);
   sequencer_->start();
   sequencer_->waitForCompletion();
-  benchmark_client_->terminate();
-  success_ = true;
-  dispatcher_->exit();
   // Save a final snapshot of the worker-specific counter accumulations before
   // we exit the thread.
   for (const auto& stat : store_.counters()) {
     // First, we strip the cluster prefix
     std::string stat_name = std::string(absl::StripPrefix(stat->name(), "cluster."));
+    stat_name = std::string(absl::StripPrefix(stat_name, "worker."));
     // Second, we strip our own prefix if it's there, else we skip.
-    const std::string worker_prefix = fmt::format("worker.{}.", worker_number_);
+    const std::string worker_prefix = fmt::format("{}.", worker_number_);
     if (stat->value() && absl::StartsWith(stat_name, worker_prefix)) {
       thread_local_counter_values_[std::string(absl::StripPrefix(stat_name, worker_prefix))] =
           stat->value();
     }
   }
+  // Note that benchmark_client_ is not terminated here, but in shutdownThread() below. This is to
+  // to prevent the shutdown artifacts from influencing the test result counters. The main thread
+  // still needs to be able to read the counters for reporting the global numbers, and those should
+  // be consistent.
 }
+
+void ClientWorkerImpl::shutdownThread() { benchmark_client_->terminate(); }
 
 StatisticPtrMap ClientWorkerImpl::statistics() const {
   StatisticPtrMap statistics;