+++ docs/root/version_history.md
@@ -1,6 +1,19 @@
 Version history
 ---------------
 
+0.3 (TBD)
+=========================
+
+### Notable (breaking) changes
+
+- In `service.proto` a change was made to allow both `Output` and `error_detail` to co-exist at the same time.
+- Both `nighthawk_client` and `nighthawk_service` will indicate execution failure (respectively through exit code or grpc reply) when connection errors and/or status code errors are observed by default.
+
+
+### Changelist
+
+- Introducing termination predicates (https://github.com/envoyproxy/nighthawk/pull/167)
+
 0.2 (July 16, 2019)
 =========================